Added stats, see description
AddScore: Added methods for adding score, which gives bonus if the player is very fast
MissedKey: Added a method for keeping track how many keys have been missed
ResetKeysMissed: Added a method for resetting the keysMissed count when farting.
GameOver: Added a method to call when the loss condition is met.